Saving data to JSON

+++ src/main/java/javaSpring/waterMeterTelegramBot/waterMeterManager/UsersManager.java
@@ -1,11 +1,13 @@
 package javaSpring.waterMeterTelegramBot.waterMeterManager;
 
+
 import javaSpring.waterMeterTelegramBot.profiles.Profile;
 import javaSpring.waterMeterTelegramBot.dataUser.User;
 import org.springframework.stereotype.Component;
 
 import java.util.HashMap;
 import java.util.Map;
+
 
 @Component
 public class UsersManager {
@@ -18,6 +20,7 @@ public class UsersManager {
         users = new HashMap<>();
     }
 
+
     ///  Создание или вызов Пользователя
     public User getUserOrCreateIfNot(String nameUser){
         User user = getUser(nameUser);
@@ -29,7 +32,7 @@ public class UsersManager {
     public User getUser(String nameUser){
         return users.get(nameUser);
     }
-    public User createUser(String nameUser, double weight){
+    public User createUser(String nameUser, int weight){
         User user = new User(nameUser, weight);
         users.put(nameUser, user);
         return user;
@@ -41,7 +44,7 @@ public class UsersManager {
     }
 
     ///  Изменение веса пользователя
-    public User setWeightToUser(String nameUser, double weight){
+    public User setWeightToUser(String nameUser, int weight){
         User user = getUser(nameUser);
         if(user != null){
             user.setWeight(weight);
